BATTLE JS functionality mostly DONE? need to add win/loss and post results
