+++ public/js/battle.js
@@ -1,5 +1,4 @@
 $(document).ready(function() {
-  var Meme = require("./memeconstruct");
   var attackMeme = {};
   var defendMeme = {};
   var battlestart = false;
@@ -7,24 +6,83 @@ $(document).ready(function() {
   var heroSelected = false;
   var defenderSelected = false;
 
-  //makes sure the game starts in its correct state
-  initBattle = () => {
-    attackMeme = {};
-    defendMeme = {};
-    battlestart = false;
-    heroSelected = false;
-    opponentSelected = false;
-  };
-  initBattle();
+  class Meme {
+    //used to construct the meme that will battle
+    constructor(id, name, lvl, ac, link, ap, hp, diceVal, boo) {
+      this.id = id;
+      this.name = name;
+      this.ac = ac;
+      this.lvl = lvl;
+      this.imgLink = link;
+      this.ap = ap;
+      this.hp = hp;
+      this.diceVal = diceVal;
+      this.isAttacker = boo;
+    }
+
+    //used to get dmg for attack
+    dmg() {
+      return this.diceroll(this.diceVal) + this.ap;
+    }
+
+    diceroll(value) {
+      var randomNumber = Math.floor(Math.random() * value) + 1;
+      return randomNumber;
+    }
+    //used to see if one hits
+    confirm(target) {
+      let confirmVal = this.diceroll(20) + this.ap;
+      console.log(`value to beat: ${target.ac}`, `confirm roll: ${confirmVal}`);
+      if (confirmVal >= target.ac) {
+        var thisAttk = this.dmg();
+        console.log(`${this.name} dealt ${thisAttk} damage!`);
+        target.hp -= thisAttk;
+      } else {
+        console.log(`${this.name} missed!`);
+      }
+    }
+
+    //used to determin how many coins and how much exp you get
+    expCoinGain(target) {
+      var lvlDiff = target.lvl - this.lvl;
+      var newValue = target.lvl * 5 + target.lvl * 2;
+      if (lvlDiff > 0) {
+        newValue += lvlDiff * 11;
+      }
+      console.log(`coins value ${newValue}`);
+      return newValue;
+      //structure this for mysql addition
+    }
+
+    concede() {
+      $.ajax({
+        type: "DESTROY",
+        url: `/api/user/memes/${this}`
+      }).then(data => {
+        //display loss modal, display total coin loss and then call update User Profile api through ajax , and then redirect to the profile page
+      });
+    }
+
+    win() {
+      //re-direct to meme page
+      $.ajax({
+        type: "PUT",
+        URL: "/api/user/id",
+        data: this.expCoinGain //win exp and coins of equal amt
+      }).then(data => {
+        //display win modal with exp and coin gains, then redirect to meme page
+      });
+    }
+  }
 
   //will run the functions need to start the battle, only if both have been selected
-  run = () => {
-    if (heroSelected === true && opponentSelected === true) {
+  var run = function() {
+    if (heroSelected === true && defenderSelected === true) {
       battlestart = true;
       postCombatant(attackMeme, "He");
       postCombatant(defendMeme, "En");
       $("#selectScreen").css("display", "none");
-      $("#battleScreen").css("display", "inline");
+      $("#battleScreen").css("display", "initial");
     } else {
       return;
     }
@@ -40,7 +98,7 @@ $(document).ready(function() {
   }
 
   function updateHP(meme, pos) {
-    $(`#hp-${pos}`).text(meme.name);
+    $(`#hp-${pos}`).text(meme.hp);
   }
 
   function fighterSelect(id) {
@@ -49,8 +107,9 @@ $(document).ready(function() {
       $.ajax(`/heros/${id}`, {
         type: "GET"
       }).then(function(meme) {
-        console.log(`attack meme ${meme}`);
+        console.log(`attack meme ${JSON.stringify(meme)}`);
         attackMeme = new Meme(
+          meme.id,
           meme.name,
           meme.lvl,
           meme.ac,
@@ -60,6 +119,8 @@ $(document).ready(function() {
           meme.dice_value,
           true
         );
+        console.log(attackMeme);
+        heroSelected = true;
         run();
       });
     } else {
@@ -73,8 +134,9 @@ $(document).ready(function() {
       $.ajax(`/combatants/${id}`, {
         type: "GET"
       }).then(function(meme) {
-        console.log(`defend meme ${meme}`);
+        console.log(`defend meme ${JSON.stringify(meme)}`);
         defendMeme = new Meme(
+          meme.id,
           meme.name,
           meme.lvl,
           meme.ac,
@@ -84,6 +146,8 @@ $(document).ready(function() {
           meme.dice_value,
           false
         );
+        console.log(defendMeme);
+        defenderSelected = true;
         run();
       });
     } else {
@@ -91,18 +155,30 @@ $(document).ready(function() {
     }
   }
 
-  $("#attackButton").on("click", () => {
+  $(".fighterSelect").on("click", function() {
+    var fighterId = $(this).attr("id");
+    fighterSelect(fighterId);
+  });
+
+  $(".defendSelect").on("click", function() {
+    var defenderId = $(this).attr("id");
+    defendSelect(defenderId);
+  });
+
+  $("#attackButton").on("click", function() {
     if (battlestart && !inTurn) {
       inTurn = true;
       attackMeme.confirm(defendMeme);
       updateHP(defendMeme, "En");
       if (defendMeme.hp <= 0) {
         attackMeme.win();
+        return;
       }
       defendMeme.confirm(attackMeme);
       updateHP(attackMeme, "He");
       if (attackMeme.hp <= 0) {
         attackMeme.concede();
+        return;
       }
 
       inTurn = false;
